prototype
Takes in a Parent Dataset which contains unique ID (key) and Name/Group ID as Attributes ( it is a JSON file), And another .xlsx which has been categorized into different sheets ("clubs"), with randomly selected unique IDs from the parent dataset along with the "interest" of that unique ID for that "club". Our code outputs a JSON file that has been populated and categorized according to different excel sheets("clubs") and gives a convenient way to store it as a database in one single JSON file.